refactor: extract common logic in a separate class
Extract common semaphore logic into ConcurrentOperationBase for reuse in concurrent handlers.
