+++ src/main/java/com/flowingcode/vaadin/addons/gridexporter/GridExporter.java
@@ -44,7 +44,7 @@ import com.vaadin.flow.server.StreamResource;
 import com.vaadin.flow.server.StreamResourceWriter;
 import com.vaadin.flow.server.VaadinSession;
 import com.vaadin.flow.server.streams.DownloadHandler;
-import com.vaadin.flow.server.streams.DownloadEvent;
+
 import com.vaadin.flow.shared.Registration;
 import java.io.Serializable;
 import java.lang.reflect.Field;
@@ -617,7 +617,8 @@ public class GridExporter<T> implements Serializable {
     }
   }
 
-  private class GridExporterConcurrentStreamResourceWriter extends ConcurrentStreamResourceWriter {
+  private class GridExporterConcurrentStreamResourceWriter extends ConcurrentStreamResourceWriter
+      implements GridExporterConcurrentStrategy {
 
     GridExporterConcurrentStreamResourceWriter(StreamResourceWriter delegate) {
       super(delegate);
@@ -626,48 +627,43 @@ public class GridExporter<T> implements Serializable {
     private Component button;
 
     @Override
+    public GridExporter<?> getExporter() {
+      return GridExporter.this;
+    }
+
+    @Override
+    public Component getButton() {
+      return button;
+    }
+
+    @Override
     public float getCost(VaadinSession session) {
-      return concurrentDownloadCost;
+      return GridExporterConcurrentStrategy.super.getCost(session);
     }
 
     @Override
     public long getTimeout() {
-        // It would have been possible to specify a different timeout for each instance but I cannot
-        // figure out a good use case for that. The timeout returned herebecomes relevant when the
-        // semaphore has been acquired by any other download, so the timeout must reflect how long
-        // it is reasonable to wait for "any other download" to complete and release the semaphore.
-      //
-        // Since the reasonable timeout would depend on the duration of "any other download", it
-      // makes sense that it's a global setting instead of a per-instance setting.
-      return GridExporterConcurrentSettings.getConcurrentDownloadTimeout(TimeUnit.NANOSECONDS);
+      return GridExporterConcurrentStrategy.super.getTimeout();
     }
 
     @Override
-    protected UI getUI() {
-      return grid.getUI().orElse(null);
+    public UI getUI() {
+      return GridExporterConcurrentStrategy.super.getUI();
     }
 
     @Override
-    protected void onTimeout() {
-      fireConcurrentDownloadTimeout();
+    public void onTimeout() {
+      GridExporterConcurrentStrategy.super.onTimeout();
     }
 
     @Override
-    protected void onAccept() {
-      if (disableOnClick) {
-        setButtonEnabled(false);
-      }
+    public void onAccept() {
+      GridExporterConcurrentStrategy.super.onAccept();
     }
 
     @Override
-    protected void onFinish() {
-      setButtonEnabled(true);
-    }
-
-    private void setButtonEnabled(boolean enabled) {
-      if (button instanceof HasEnabled) {
-        grid.getUI().ifPresent(ui -> ui.access(() -> ((HasEnabled) button).setEnabled(enabled)));
-      }
+    public void onFinish() {
+      GridExporterConcurrentStrategy.super.onFinish();
     }
   }
 
@@ -678,7 +674,8 @@ public class GridExporter<T> implements Serializable {
    * GridExporterConcurrentStreamResourceWriter
    * but for the new API.
    */
-  private class GridExporterConcurrentDownloadHandler extends ConcurrentDownloadHandler {
+  public class GridExporterConcurrentDownloadHandler extends ConcurrentDownloadHandler
+      implements GridExporterConcurrentStrategy {
 
     GridExporterConcurrentDownloadHandler(DownloadHandler delegate) {
       super(delegate);
@@ -687,41 +684,43 @@ public class GridExporter<T> implements Serializable {
     private Component button;
 
     @Override
+    public GridExporter<?> getExporter() {
+      return GridExporter.this;
+    }
+
+    @Override
+    public Component getButton() {
+      return button;
+    }
+
+    @Override
     public float getCost(VaadinSession session) {
-      return concurrentDownloadCost;
+      return GridExporterConcurrentStrategy.super.getCost(session);
     }
 
     @Override
     public long getTimeout() {
-      return GridExporterConcurrentSettings.getConcurrentDownloadTimeout(TimeUnit.NANOSECONDS);
+      return GridExporterConcurrentStrategy.super.getTimeout();
     }
 
     @Override
-    protected UI getUI() {
-      return grid.getUI().orElse(null);
+    public UI getUI() {
+      return GridExporterConcurrentStrategy.super.getUI();
     }
 
     @Override
-    protected void onTimeout() {
-      fireConcurrentDownloadTimeout();
+    public void onTimeout() {
+      GridExporterConcurrentStrategy.super.onTimeout();
     }
 
     @Override
-    protected void onAccept() {
-      if (disableOnClick) {
-        setButtonEnabled(false);
-      }
+    public void onAccept() {
+      GridExporterConcurrentStrategy.super.onAccept();
     }
 
     @Override
-    protected void onFinish() {
-      setButtonEnabled(true);
-    }
-
-    private void setButtonEnabled(boolean enabled) {
-      if (button instanceof HasEnabled) {
-        grid.getUI().ifPresent(ui -> ui.access(() -> ((HasEnabled) button).setEnabled(enabled)));
-      }
+    public void onFinish() {
+      GridExporterConcurrentStrategy.super.onFinish();
     }
 
     /**
@@ -1142,6 +1141,45 @@ public class GridExporter<T> implements Serializable {
 
   public void setCsvCharset(SerializableSupplier<Charset> charset) {
     csvCharset = charset;
+  }
+
+  private interface GridExporterConcurrentStrategy {
+    GridExporter<?> getExporter();
+
+    Component getButton();
+
+    default float getCost(VaadinSession session) {
+      return getExporter().concurrentDownloadCost;
+    }
+
+    default long getTimeout() {
+      return GridExporterConcurrentSettings.getConcurrentDownloadTimeout(TimeUnit.NANOSECONDS);
+    }
+
+    default UI getUI() {
+      return getExporter().grid.getUI().orElse(null);
+    }
+
+    default void onTimeout() {
+      getExporter().fireConcurrentDownloadTimeout();
+    }
+
+    default void onAccept() {
+      if (getExporter().disableOnClick) {
+        setButtonEnabled(false);
+      }
+    }
+
+    default void onFinish() {
+      setButtonEnabled(true);
+    }
+
+    default void setButtonEnabled(boolean enabled) {
+      Component button = getButton();
+      if (button instanceof HasEnabled) {
+        getExporter().grid.getUI().ifPresent(ui -> ui.access(() -> ((HasEnabled) button).setEnabled(enabled)));
+      }
+    }
   }
 
 }
